@non-Jedi 2019-09-07
Very cool. I think we can probably polish it to be more idiomatic Julia. My preference would be to just use NTuples with unsafe_store! for now. I think getting AOT compilation working consistently on the benchmarks-game machine might be a ways off. And it might not be accepted at all depending on whether the maintainer wants to deal with the headache.